feat(RHTAPREL-661): allow Docker's golang image
To address the issue describe here [1], controller-runtime
needs to be upgraded in RHTAP operators. A consequence of this
is that Go has to be upgraded to 1.20 as well. The current builder,
go-toolset, will nto have support for this version until two weeks
from now with the release of RHEL 9.3. Until then, we need to rely
on Docker's golang image.

Signed-off-by: David Moreno García <damoreno@redhat.com>