Remove irrelevant step image
As the comment suggests, the
quay.io/openshift-release-dev/ocp-v4.0-art-dev was added as a workaround
to an OCP issue. The policy rule that uses this data has been moved from
the "release" namespace to the "task" namespace. This means the policy
rule applies to a Task definition instead of the Task information
captured in the SLSA Provenance attestation. The check is now executed
as part of the build-definitions CI on the Task definitions directly.

At some point, we may want to execute the check against the information
in the SLSA Provenance attestation but currently this is not possible.
https://issues.redhat.com/browse/RFE-4608 will hopefully turn this into
reality.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>