Allow usage of the *-oci-ta Task variants
As part of EC-20, we are adding Task variants that support Trusted
Artifacts with OCI storage. These have the suffix of `-oci-ta`. These
Tasks function similarly to their counter-parts. The main difference is
that they share data between them via resources in an OCI registry
instead of file in a shared PVC on-cluster.

This commit allows the `-oci-ta` Tasks to fullfill the required-tasks
check.

Additionally, the `summary` Task is no longer marked as required. This
was the outcome after some investigation on its actual value. More
context is available on the comments of EC-643.

Finally, older entries that are no longer relevant have been removed
from the list.

Ref: EC-644

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>