Make ecosytem preflight failures non-blocking
I'm not sure about the pros and cons of this, but there are several
teams who get surprised by this failure and want to exclude it.

Because the task itself it not included in the required task lists,
removing it entirely from the pipeline gets EC to pass. So for that
reason I'm assuming it's acceptable to make it non-blocking.

(The downside is that it's going to be much easier to ignore any
failures produced by this test.)