Drop docs
Remove the antora extension that generates the documentation for the
data files defined in this repository.

There are a couple of reasons for this.

First, it is our understanding that the webpage that uses the antora
extension is rarely used. A quick sample of some users reveal that they
weren't even aware of its existence.

Second, this is very specific to Konflux and arugably does not really
belong in the upstream EC website.

https://github.com/enterprise-contract/ec-cli/issues/1810 has been
created to make the output of `ec inspect policy-data` a little more
informative which could be useful for others.

Ref: EC-587

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>