Require specific value for vendor label
Currently, the EC CLI behaves differently if the vendor label is set to
the value `Red Hat, Inc.`. If so, it downloads certain files from within
the image and makes them available for processing by the rego policy
rules. If the image uses a different value for this label then those
files are not available. This often causes confusion to users and to
most of us analyzing those failures. This change increases visibility
for such cases.

Ref: https://issues.redhat.com/browse/EC-787

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>