Drop rpms-signature-scan from fbc required tasks
FBC images are "catalog fragments" that are never actually shipped.
They only exist to convey data to the process that updates the global
index image. It is unreasonable to scan them for rpm content.

Even if they somehow included unsigned rpm content, that content is
never going to be exposed.