Add three new sast scans to required tasks
Use an effective on date of March 1st to give people time to get
them added them to their pipeline definitions.

Note the tests will be required, meaning if they're not included EC
will produce a "required task not found" violation, but they are
"informative" only, meaning if they produce a failure status, EC
will not produce a violation.

I didn't add the new tasks to the fbc pipeline required task list
since I think sast scans are likely not useful there, and because
the existing sast-snyk-check scan is already not required for fbc.

Handy queries to help confirm this change:
  yq '.pipeline-required-tasks | to_entries[].value |[.[0].tasks - .[1].tasks]' data/required_tasks.yml
  yq '.required-tasks|[.[0].tasks - .[1].tasks]' data/required_tasks.yml

Ref: https://issues.redhat.com/browse/EC-887
Ref: https://issues.redhat.com/browse/EC-888
Ref: https://issues.redhat.com/browse/EC-889