Extend deadline for new sast tasks
The work to generate PRs to add the new SAST tasks is behind
schedule. See EC-1063 for more details.

It's generally working but it's needs a bit more work and testing
before it's ready for production use. Also I'd like to have the
email notifications and announcements sent, and the generated PRs
created, well ahead of the deadline, to give teams have enough time
to respond, and so we have enough time to deal with edge cases and
unexpected problems.

Ref: https://issues.redhat.com/browse/EC-1063