Permit inheritance of cpe id again
It turns out that this rule doesn't actually check if the value was
implicitly inherited. It only checks if the value is the same or
different between parent and child. This means that if you legitimately
have a parent and child with the same cpe id, and even if you explicitly
set those labels on parent and child, conforma will still fail because
the value is the same between parent and child.

That's not acceptable behavior, so - removing this from the rules.

We will have to depend on features in the release pipelines, instead:

* https://github.com/konflux-ci/release-service-catalog/pull/1615
* https://github.com/konflux-ci/release-service-catalog/pull/1657
* https://github.com/konflux-ci/release-service-catalog/pull/1681